visualizaciones e historial (hecho con Jose)

+++ src/main/java/com/unicast/unicast_backend/persistance/model/Display.java
@@ -31,7 +31,7 @@ public class Display {
 	
     private Timestamp timestampLastTime;
     
-    private Long secsFromBeg;
+    private Integer secsFromBeg;
 	
 	
 }
